+++ sms_frame_yeastar/__openerp__.py
@@ -1,6 +1,6 @@
 {
     'name': "SMS Framework - Yeastar TG 100",
-    'version': "1.1",
+    'version': "1.1.2",
     'author': "Sythil Tech",
     'category': "Tools",
     'support': "steven@sythiltech.com.au",