+++ sms_frame_yeastar/models/__init__.py
@@ -1,1 +1,1 @@
-import sms_gateway_yeastar
+import sms_gateway_yeastar, res_partner