+++ sms_frame_yeastar/doc/changelog.rst
@@ -1,3 +1,11 @@
+v1.1.2
+======
+* Map received sms to partner record
+
+v1.1.1
+======
+* Fix receive SMS database commit issue
+
 v1.1
 ====
 * Basic receive SMS